requirements.txt
# YOLO Detection System - Basler USB3 Vision
# Requirements para Windows com CUDA 12.x

# PyTorch com CUDA 12.1 (instalar separadamente)
# pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu121

torch>=2.0.0
torchvision>=0.15.0

# Ultralytics YOLO
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Basler Pylon
pypylon>=2.0.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Configuration
PyYAML>=6.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0